Separate vendor detection to a library package
